add: move variable
