+++ src/app/prompts/[id]/page.tsx
@@ -4,9 +4,8 @@ import {ReactElement} from "react";
 import {EntityNotFoundError} from "typeorm";
 import NotFound from "next/dist/client/components/not-found-error";
 import {db} from "@/lib/data";
-import {Prompt} from "@/db/entity/prompt";
+import {Prompt, PromptVariable} from "@/db/entity/prompt";
 import {Form, State} from "@/ui/form";
-import {PromptVariable} from "@/db/entity/prompt_variable";
 import Link from "next/link";
