+++ src/db/entity/prompt.ts
@@ -2,8 +2,39 @@
 
 import "reflect-metadata";
 import type {Relation} from "typeorm";
-import {Column, Entity, OneToMany, PrimaryGeneratedColumn} from "typeorm";
-import {PromptVariable} from "@/db/entity/prompt_variable";
+import {Column, Entity, JoinColumn, ManyToOne, OneToMany, PrimaryGeneratedColumn} from "typeorm";
+
+@Entity({name: 'prompt_variables'})
+export class PromptVariable {
+
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @Column({
+        type: "varchar",
+        length: 32
+    })
+    name: string;
+
+    @Column({
+        type: "varchar",
+        length: 32
+    })
+    type: string;
+
+    @ManyToOne(type => Prompt, prompt => prompt.promptVariables)
+    @JoinColumn(
+        {name: "prompt_id"}
+    )
+    prompt: Relation<Prompt>;
+
+    @Column({
+        default: true,
+        name: "is_active"
+    })
+    isActive: boolean;
+
+}
 
 @Entity({name: 'prompts'})
 export class Prompt {
